Set up user w/ name and password in test container
Create a non-privileged user for testing via `docker exec` after booting
a container from the "sshkit-test-sshd" image.

Setting the user password requires an indirection because `chpasswd`
reads password input from STDIN and using pipes causes problems when
we're actually trying to run and pipe a command inside the container.

Quoting is not an option, since `docker exec` does not accept quoted
commands. For illustration:

  docker exec 09abc3f5 echo user:pass | chpasswd
  # your local shell will process the pipe, not what you want

  docker exec 'echo user:pass | chasswd'
  # is not accepted by docker

As a solution, what we do is:

  docker exec sh -c 'echo user:pass | chpasswd'